+++ app/server/models/VideoModel.js
@@ -388,6 +388,7 @@ VideoModel.prototype.add = function(media, callback) {
     id: String(media.id),
     available: media.available,
     title: media.title,
+    shortDescription: media.shortDescription,
     description: media.description,
     state: media.state,
     date: media.date,
@@ -964,6 +965,8 @@ VideoModel.prototype.update = function(id, data, callback) {
     info.title = data.title;
   if (data.date)
     info.date = data.date;
+  if (data.shortDescription)
+    info.shortDescription = data.shortDescription;
   if (data.description)
     info.description = data.description;
   if (data.properties)